forensic_ediscovery/requirements.txt
# Forensic E-Discovery Email Processing System
# Core dependencies

# Gmail API
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.114.0

# Gemini AI SDK
google-genai>=1.0.0

# Data validation and schemas
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Email parsing
email-validator>=2.1.0

# File handling
python-magic>=0.4.27;sys_platform!="win32"
python-magic-bin>=0.4.14;sys_platform=="win32"

# CLI and utilities
click>=8.1.0
rich>=13.7.0
python-dotenv>=1.0.0

# Progress and async
tqdm>=4.66.0
aiofiles>=23.2.0

# Date handling
python-dateutil>=2.8.0
